Fix 'operator +=' is ambiguous
Fixes this error:

spark\demos\src\SPKTestIrrlicht.cpp(168,129): error C2593: 'operator +=' is ambiguous
irrlicht\include\irrString.h(1053,20): message : could be 'irr::core::string<wchar_t,irr::core::irrAllocator<T>> &irr::core::stri
ng<T,irr::core::irrAllocator<T>>::operator +=(const float)'
          with
          [
              T=wchar_t
          ]

+++ demos/src/SPKTestIrrlicht.cpp
@@ -165,7 +165,7 @@ int main(int argc, char *argv[])
 		// Renders scene
 		smgr->drawAll();
 
-		irr::core::stringw infos; infos+="FPS: "; infos+=driver->getFPS(); infos+=" - Nb Particles: "; infos+=system->getNbParticles();
+		irr::core::stringw infos; infos+="FPS: "; infos+=driver->getFPS(); infos+=" - Nb Particles: "; infos+=static_cast<unsigned int>(system->getNbParticles());
         guienv->getBuiltInFont()->draw(infos.c_str(),irr::core::rect<irr::s32>(0,0,170,20),irr::video::SColor(255,255,255,255));
 
 		driver->endScene();
